+++ test/functional-portable/dynamicrp/noncloud/resources/testdata/usertypealpha.yaml
@@ -16,3 +16,36 @@ types:
           required:
             - application
             - environment
+  postgres:
+    apiVersions:
+      '2023-10-01-preview':
+        schema: 
+          type: object
+          properties:
+            environment:
+              type: string
+            application:
+              type: string
+              description: The resource ID of the application.
+            status:
+              type: object
+              properties:
+                binding:
+                  type: object
+                  properties:
+                    database:
+                      type: string
+                      description: The name of the database.
+                    host:
+                      type: string
+                      description: The host name of the database.
+                    port:
+                      type: string
+                      description: The port number of the database.
+                    username:
+                      type: string
+                      description: The username for the database.
+                    password:
+                      type: string
+                      description: The password for the database.
+    capabilities: ["SupportsRecipes"]
